+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.0.1 (21 Aug 2019)
+
+Bugfixes
+- InputStream: getSize returns the content of the `$_SERVER['CONTENT_LENGTH']` variable. 
+- OutputStream: getSize throws a exception since the size can not be determined.
+
 ## 1.0.0 (21 Aug 2019)
 
 Initial release